feat: enable shell error checks
Enable error checks to:

- Exit on errors
- Disallow empty variables
- Fail when a piped command errors

+++ lib/linter.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/env bash
 
+set -o errexit
+set -o nounset
+set -o pipefail
+
 ##################################################################
 # Debug Vars                                                     #
 # Define these early, so we can use debug logging ASAP if needed #
@@ -180,6 +184,7 @@ debug "TFLINT_LOG: ${TFLINT_LOG}"
 ANSIBLE_FILE_NAME="${ANSIBLE_CONFIG_FILE:-.ansible-lint.yml}"
 # shellcheck disable=SC2034  # Variable is referenced indirectly
 ARM_FILE_NAME=".arm-ttk.psd1"
+BASH_SEVERITY="${BASH_SEVERITY:-""}"
 CHECKOV_FILE_NAME="${CHECKOV_FILE_NAME:-".checkov.yaml"}"
 # shellcheck disable=SC2034  # Variable is referenced indirectly
 CLOJURE_FILE_NAME=".clj-kondo/config.edn"
@@ -224,6 +229,7 @@ KUBERNETES_KUBECONFORM_OPTIONS="${KUBERNETES_KUBECONFORM_OPTIONS:-null}"
 LATEX_FILE_NAME=".chktexrc"
 # shellcheck disable=SC2034  # Variable is referenced indirectly
 LUA_FILE_NAME=".luacheckrc"
+MARKDOWN_CUSTOM_RULE_GLOBS="${MARKDOWN_CUSTOM_RULE_GLOBS:-""}"
 # shellcheck disable=SC2034  # Variable is referenced indirectly
 MARKDOWN_FILE_NAME="${MARKDOWN_CONFIG_FILE:-.markdown-lint.yml}"
 # shellcheck disable=SC2034  # Variable is referenced indirectly
@@ -419,31 +425,21 @@ LINTER_NAMES_ARRAY['YAML']="yamllint"
 ##########################
 for LANGUAGE in "${LANGUAGE_ARRAY[@]}"; do
   FILE_ARRAY_VARIABLE_NAME="FILE_ARRAY_${LANGUAGE}"
-  debug "Setting ${FILE_ARRAY_VARIABLE_NAME} variable..."
+  debug "Initializing ${FILE_ARRAY_VARIABLE_NAME}"
   eval "${FILE_ARRAY_VARIABLE_NAME}=()"
 done
 
-################################################################################
-########################## FUNCTIONS BELOW #####################################
-################################################################################
-################################################################################
-#### Function Header ###########################################################
 Header() {
-  ###############################
-  # Give them the possum action #
-  ###############################
   if [[ "${SUPPRESS_POSSUM}" == "false" ]]; then
     /bin/bash /action/lib/functions/possum.sh
   fi
 
-  ##########
-  # Prints #
-  ##########
   info "---------------------------------------------"
   info "--- GitHub Actions Multi Language Linter ----"
-  info " - Image Creation Date:[${BUILD_DATE}]"
-  info " - Image Revision:[${BUILD_REVISION}]"
-  info " - Image Version:[${BUILD_VERSION}]"
+  # These variable may be not set when we run the script to generate the versions file
+  info " - Image Creation Date: ${BUILD_DATE:-"not set"}"
+  info " - Image Revision: ${BUILD_REVISION:-"not set"}"
+  info " - Image Version: ${BUILD_VERSION:-"not set"}"
   info "---------------------------------------------"
   info "---------------------------------------------"
   info "The Super-Linter source code can be found at:"
@@ -469,19 +465,14 @@ ConfigureGitSafeDirectories() {
   done
 }
 
-################################################################################
-#### Function GetGitHubVars ####################################################
 GetGitHubVars() {
-  ##########
-  # Prints #
-  ##########
   info "--------------------------------------------"
   info "Gathering GitHub information..."
 
   if [[ ${RUN_LOCAL} != "false" ]]; then
     info "RUN_LOCAL has been set to: ${RUN_LOCAL}. Bypassing GitHub Actions variables..."
 
-    if [ -z "${GITHUB_WORKSPACE}" ]; then
+    if [ -z "${GITHUB_WORKSPACE:-}" ]; then
       GITHUB_WORKSPACE="${DEFAULT_WORKSPACE}"
     fi
 
@@ -616,12 +607,8 @@ GetGitHubVars() {
   # We need this for parallel
   export GITHUB_WORKSPACE
 }
-################################################################################
-#### Function CallStatusAPI ####################################################
+
 CallStatusAPI() {
-  ####################
-  # Pull in the vars #
-  ####################
   LANGUAGE="${1}" # language that was validated
   STATUS="${2}"   # success | error
   SUCCESS_MSG='No errors were found in the linting process'
@@ -728,8 +715,7 @@ Footer() {
 
   exit ${SUPER_LINTER_EXIT_CODE}
 }
-################################################################################
-#### Function UpdateLoopsForImage ##############################################
+
 UpdateLoopsForImage() {
   ######################################################################
   # Need to clean the array lists of the linters removed for the image #
@@ -769,36 +755,36 @@ UpdateLoopsForImage() {
 cleanup() {
   local -ri EXIT_CODE=$?
 
-  debug "Removing temporary files and directories"
-  rm -rfv \
-    "${GITHUB_WORKSPACE}/logback.log"
+  if [ -n "${GITHUB_WORKSPACE:-}" ]; then
+    debug "Removing temporary files and directories"
+    rm -rfv \
+      "${GITHUB_WORKSPACE}/logback.log"
 
-  if [ "${SUPER_LINTER_COPIED_R_LINTER_RULES_FILE}" == "true" ]; then
-    debug "Deleting ${R_RULES_FILE_PATH_IN_ROOT} because super-linter created it."
-    rm -rfv "${R_RULES_FILE_PATH_IN_ROOT}"
-  fi
+    if [[ "${SUPER_LINTER_COPIED_R_LINTER_RULES_FILE:-}" == "true" ]]; then
+      debug "Deleting ${R_RULES_FILE_PATH_IN_ROOT} because super-linter created it."
+      rm -rfv "${R_RULES_FILE_PATH_IN_ROOT}"
+    fi
 
-  # Define this variable here so we can rely on it as soon as possible
-  local LOG_FILE_PATH="${GITHUB_WORKSPACE}/${LOG_FILE}"
-  debug "LOG_FILE_PATH: ${LOG_FILE_PATH}"
-  if [ "${CREATE_LOG_FILE}" = "true" ]; then
-    debug "Moving log file from ${LOG_TEMP} to ${LOG_FILE_PATH}"
-    mv \
-      --force \
-      --verbose \
-      "${LOG_TEMP}" "${LOG_FILE_PATH}"
+    # Define this variable here so we can rely on it as soon as possible
+    local LOG_FILE_PATH="${GITHUB_WORKSPACE}/${LOG_FILE}"
+    debug "LOG_FILE_PATH: ${LOG_FILE_PATH}"
+    if [ "${CREATE_LOG_FILE}" = "true" ]; then
+      debug "Moving log file from ${LOG_TEMP} to ${LOG_FILE_PATH}"
+      mv \
+        --force \
+        --verbose \
+        "${LOG_TEMP}" "${LOG_FILE_PATH}"
+    else
+      debug "Skipping the moving of the log file from ${LOG_TEMP} to ${LOG_FILE_PATH}"
+    fi
   else
-    debug "Skipping the moving of the log file from ${LOG_TEMP} to ${LOG_FILE_PATH}"
+    debug "GITHUB_WORKSPACE is not set. Skipping filesystem cleanup steps"
   fi
 
   exit "${EXIT_CODE}"
   trap - 0 1 2 3 6 14 15
 }
 trap 'cleanup' 0 1 2 3 6 14 15
-
-################################################################################
-############################### MAIN ###########################################
-################################################################################
 
 ##########
 # Header #
